+++ framework/base/src/types/managed/basic/big_float.rs
@@ -35,6 +35,14 @@ impl<M: ManagedTypeApi> ManagedType<M> for BigFloat<M> {
         self.handle.clone()
     }
 
+    unsafe fn forget_into_handle(self) -> Self::OwnHandle {
+        unsafe {
+            let handle = core::ptr::read(&self.handle);
+            core::mem::forget(self);
+            handle
+        }
+    }
+
     fn transmute_from_handle_ref(handle_ref: &M::BigFloatHandle) -> &Self {
         unsafe { core::mem::transmute(handle_ref) }
     }