+++ framework/base/src/types/managed/wrapped/big_uint.rs
@@ -36,6 +36,10 @@ impl<M: ManagedTypeApi> ManagedType<M> for BigUint<M> {
         self.value.handle.clone()
     }
 
+    unsafe fn forget_into_handle(self) -> Self::OwnHandle {
+        self.value.forget_into_handle()
+    }
+
     fn transmute_from_handle_ref(handle_ref: &M::BigIntHandle) -> &Self {
         unsafe { core::mem::transmute(handle_ref) }
     }